+++ vite.config.ts
@@ -1,5 +1,6 @@
 import { defineConfig } from 'vite';
 import dts from 'vite-plugin-dts';
+import svgr from 'vite-plugin-svgr';
 import { name } from './package.json';
 
 export default defineConfig({
@@ -21,5 +22,5 @@ export default defineConfig({
     sourcemap: true,
     emptyOutDir: true,
   },
-  plugins: [dts({ tsconfigPath: 'tsconfig.build.json' })],
+  plugins: [dts({ tsconfigPath: 'tsconfig.build.json' }), svgr()],
 });